Add support for creating merge commits to Git::PurePerl::Object::Commit
Git::PurePerl::Object::Commit 'parent' was a simple Str, and content
was built with a single 'parent' line. This meant it was impossible
to create merge commits (commits with multiple parents).

This patch changes the 'isa' constraint for 'parent', making it
possible to pass a reference to an array of parents.

_build_content is modified to behave correctly for all value types.